backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<1.9.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.23,<1.5.0
psycopg2-binary>=2.9.1,<2.10.0
passlib>=1.7.4,<1.8.0
python-jose>=3.3.0,<3.4.0
python-multipart>=0.0.5,<0.1.0
influxdb-client>=1.24.0,<1.25.0
httpx>=0.18.2,<0.19.0
pytest>=6.2.5,<6.3.0
python-dotenv>=0.19.0,<0.20.0
websockets>=10.0,<11.0
redis>=4.0.2,<4.1.0
yfinance>=0.1.63,<0.2.0
pandas>=1.3.2,<1.4.0
textblob>=0.17.1,<0.18.0
alembic>=1.7.3,<1.8.0
bcrypt>=3.2.0,<3.3.0
email-validator>=1.1.3,<1.2.0
aioredis>=2.0.0,<2.1.0
asyncpg>=0.24.0,<0.25.0